Edits for Pull Request #44

+++ Software/Arduino code/Sidereal.cpp
@@ -3,11 +3,11 @@
 static DayTime Sidereal::calculateByGPS(TinyGPSPlus* gps){
     DayTime timeUTC = DayTime(gps->time.hour(), gps->time.minute(), gps->time.second());
     int deltaJd = calculateDeltaJd(gps->date.year(), gps->date.month(), gps->date.day());
-    float deltaJ = (float) deltaJd + ((timeUTC.getTotalHours()) / 24.0);
-    return DayTime((float) (calculateTheta(deltaJ, (float) gps->location.lng(), timeUTC.getTotalHours()) / 15.0));
+    double deltaJ = deltaJd + ((timeUTC.getTotalHours()) / 24.0);
+    return DayTime((float)(calculateTheta(deltaJ, gps->location.lng(), timeUTC.getTotalHours()) / 15.0));
 }
 
-static const float Sidereal::calculateTheta(float deltaJ, float longitude, float timeUTC){
+static const double Sidereal::calculateTheta(double deltaJ, double longitude, float timeUTC){
     float theta = C1;
     theta += C2 * deltaJ;
     theta += C3 * timeUTC;
@@ -20,10 +20,14 @@ static const int Sidereal::calculateDeltaJd(int year, int month, int day){
 
     // Calculating days without leapdays
     long int deltaJd = (year - REFERENCE_JEAR) * 365 + day; 
-    for (int i=0; i < month - 1; i++){deltaJd += daysInMonth[i];}
+    for (int i=0; i < month - 1; i++){
+        deltaJd += daysInMonth[i];
+    }
 
     // Add leapdays
-    if (month <= 2){year--;}
+    if (month <= 2){
+        year--;
+    }
     deltaJd += (year / 4 - year / 100 + year / 400);
     deltaJd -= (REFERENCE_JEAR / 4 - REFERENCE_JEAR / 100 + REFERENCE_JEAR / 400);
     return deltaJd;
